Heavy Refactor
Heavily refactored entire project.

-Moved entire ui drawing and appearance to kv file.
-Added a ui logic file to handle communicating between lesson logic and the ui.
-UI now no longer makes calls directly to lesson logic.
-Refactored and removed some old methods that were bad or no longer in use.
-Added ids to profiles and result layout buttons
-disabled next lesson if there are no more lessons on result screen
-added profile logic, doesn't do anything right now, but the file is there.

Factoring is finished. Can conclude I am finished with this milestone.

Next up will be Questions, Answers, and the Lessons Implementation. :)
