+++ profileLogic.py
@@ -1,0 +1,7 @@
+profileNames = ["Amanda", "Brian", "Mike", "Claire", "Sophie", "Jeff"]
+
+def getProfileNameLength():
+    return len(profileNames)
+
+def getProfileName(index):
+    return profileNames[index]
